requirements.txt
streamlit
pandas
langchain>=0.1.12
langchain-openai>=0.0.8
langchain-community>=0.0.28
python-dotenv
pydantic
chromadb
sentence-transformers
plotly
matplotlib
reportlab
fpdf2
pytest
fastapi
uvicorn
python-multipart
langchain-google-genai
sqlalchemy
psycopg2-binary
